% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/posts.R
\name{getposts}
\alias{getposts}
\title{getposts}
\usage{
getposts(forum = forum, thread = thread, ...)
}
\arguments{
\item{forum}{string or numeric}

\item{thread}{string presumably collected by getthreadid.}

\item{...}{addition listTemplate options.}
}
\description{
Wrapper for posts to extract and clean the returned posts.
}

