% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{getthreadid}
\alias{getthreadid}
\title{getthreadid}
\usage{
getthreadid(thrd = NULL, ...)
}
\arguments{
\item{thrd}{numeric or string if is url this wrapper tries to fetch the id of
the disqus thread in the link.}

\item{...}{additional listTemplate Options.}
}
\description{
Wrapper for threads() to fetch the id of a given forum.
}
