% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{posts}
\title{Disqus Posts API}
\usage{
posts(...)
}
\arguments{
\item{option}{string close, create, details, list, listHot, listPopular,
listPosts, open, remove, restore, set, subscribe, unsubscribe, update, vote
Options may differ in their set of Arguments.}

\item{category}{integer}

\item{forum}{integer/string unique identifier of a forum}

\item{thread}{integer or string if integer must be unique if string must
have link: or ident:}

\item{author}{}

\item{since}{integer or string: unixtime or iso timestamp
\code{format(Sys.time(), "\%Y-\%m-\%dT\%H:\%M:\%S")}.}

\item{related}{}

\item{cursor}{}

\item{attach}{}

\item{limit}{integer between 25 (default) and 100 (maximum)}

\item{include}{}

\item{order}{string "desc" (descending) or "asc" (ascending)}

\item{format}{string json, jsonp or rss}
}
\description{
Disqus Posts API
}
\details{
Original API Documentation for threads
\code{https://disqus.com/api/docs/threads/}
}
\examples{
# lists threads of forum=politico in RSS-format
threads("list", forum="politico", format="rss")

# use a specific link ending with ".html" (may work if link was not changed)
art <- "http://www.politico.com/magazine/story/2015/05/fox-news-liberals-118235.html"
arts <- threads("list" , forum="politico", thread=paste0("link:", art))
postslist <- posts("list", thread=arts[1,"id"])
# use a timestamp
unixtime <- as.numeric(as.POSIXct("2015-05-26", format="\%Y-\%m-\%d"))
threads("list", forum="politico", since=unixtime)
}

