% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postTemplate.R
\name{postTemplate}
\alias{postTemplate}
\title{postTemplate}
\usage{
postTemplate(option = NULL, title = NULL, message = NULL,
  category = NULL, forum = NULL, thread = NULL, author = NULL,
  since = NULL, related = NULL, cursor = NULL, attach = NULL,
  limit = 25, include = NULL, order = NULL, format = "json",
  ressource = c("threads", "posts"), atoken = NULL, pubkey, seckey)
}
\arguments{
\item{option}{option}

\item{title}{title}

\item{message}{message}

\item{category}{category}

\item{forum}{forum}

\item{thread}{thread}

\item{author}{author}

\item{since}{since}

\item{related}{related}

\item{cursor}{cursor}

\item{attach}{attach}

\item{limit}{limit}

\item{include}{include}

\item{order}{order}

\item{format}{string default json}

\item{ressource}{ressource}

\item{atoken}{atorken}

\item{pubkey}{pubkey}

\item{seckey}{seckey}
}
\description{
postTemplate
}
