% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{threads}
\alias{threads}
\title{Disqus Threads API}
\usage{
threads(option, type = "json", limit = 25, ...)
}
\arguments{
\item{option}{option default NULL}

\item{type}{json or rss default json}

\item{limit}{integer limit 1 to 100}

\item{...}{options: close, create, details, list, listHot, listPopular,
listPosts, open, remove, restore, set, subscribe, unsubscribe, update, vote

GET functions:
details
list
listHot
listPopular
listPosts
set
subscribe
update

POST functions: close,
create
open
remove
restore
unsubscribe
vote -1, 0, 1}
}
\description{
\code{threads()} is a function to access the disqus threads api. It is built
as a comfortable wrapper for \code{listTemplate()} a disqusR function
underlying most of the retrieving (GET) functions of the disqus API. It will
return an object of class dq_thread, containing the content retrieved.
By default \code{threads()} will return a json object.
}
\details{
threads() requires at least one option.

For a full documentation and all currently possible options please see
\url{https://disqus.com/api/docs/threads/}.

As of June 2015 following options are available.
}
\examples{
\dontrun{
# lists threads of forum=politico in RSS-type
threads("list", forum="politico", type="rss")

# list a few disqus threads
threads("list")
}
}
