% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disqusR.R
\docType{package}
\name{disqusR}
\alias{disqusR}
\alias{disqusR-package}
\title{Access to the version 3.0 Disqus API.}
\description{
Check applications(), blacklist(), category(), exports(), forums(),
imports(), threads(), topics(), trends(), users() and whitelist().
}
\note{
If you catch a bug, please do not sue us, we do not have any money.
}
\author{
Marvin Garbuszus \email{jan.garbuszus@ruhr-uni-bochum.de}

Sebastian Jeworutzki \email{sebastian.jeworutzki@ruhr-uni-bochum.de}
}
