% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{posts}
\title{Disqus Posts API}
\usage{
posts(option, type = "json", limit = 25, ...)
}
\arguments{
\item{...}{listTemplate option.}
}
\description{
Disqus Posts API
}
\details{
Original API Documentation for threads
\code{https://disqus.com/api/docs/threads/}
}
\examples{
\dontrun{
# lists threads of forum=politico in RSS-format
threads("list", forum="politico", format="rss")

# use a specific link ending with ".html" (may work if link was not changed)
art <- "http://www.11freunde.de/liveticker/das-rueckspiel-der-relegation-im-liveticker"
thrd <- threads("list", thread=art)
pst <- posts("list", forum="11freunde", thread=thrd$id)

# use a timestamp
unixtime <- as.numeric(as.POSIXct("2015-05-26", format="\%Y-\%m-\%d"))
threads("list", forum="politico", since=unixtime)
}
}
