% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posts.R
\name{posts}
\alias{posts}
\title{Disqus Posts API}
\usage{
posts(...)
}
\arguments{
\item{...}{listTemplate option.}

\item{category}{integer}

\item{forum}{integer/string unique identifier of a forum}

\item{thread}{integer or string if integer must be unique if string must
have link: or ident:}

\item{since}{integer or string: unixtime or iso timestamp
\code{format(Sys.time(), "\%Y-\%m-\%dT\%H:\%M:\%S")}.}

\item{limit}{integer between 25 (default) and 100 (maximum)}

\item{order}{string "desc" (descending) or "asc" (ascending)}

\item{format}{string json, jsonp or rss}
}
\description{
Disqus Posts API
}
\details{
Original API Documentation for threads
\code{https://disqus.com/api/docs/threads/}
}
\examples{
# lists threads of forum=politico in RSS-format
threads("list", forum="politico", format="rss")

# use a specific link ending with ".html" (may work if link was not changed)
art <- "http://www.11freunde.de/liveticker/das-rueckspiel-der-relegation-im-liveticker"
thrd <- threads("list", thread=art)
pst <- posts("list", forum="11freunde", thread=thrd$id)

# use a timestamp
unixtime <- as.numeric(as.POSIXct("2015-05-26", format="\%Y-\%m-\%d"))
threads("list", forum="politico", since=unixtime)

}
